+++ src/main/java/com/example/testiit/repository/HistoryRepository.java
@@ -3,5 +3,5 @@ package com.example.testiit.repository;
 import com.example.testiit.model.CoffeeMakerHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoryRepository extends JpaRepository<CoffeeMakerHistory,Integer> {
+public interface HistoryRepository extends JpaRepository<CoffeeMakerHistory,Long> {
 }
